Improved structure of the projects to individual services
